Adding GFileMonitors to media-server.
So that media folder changes are reported instantly to apps using it
(gallery, music), on both internal and sd card storage.

Prior to the this patch, when a new media was added, in order to see
it in the respective app, the user had to reboot the phone. This
problem happens because the media server did not monitor the
/opt/media directory, it would only index media files during phone
startup.

To fix this problem, media server will now use GFileMonitors to watch
the /opt/media directory and, if a new file is added to this
directory, it will know and it will reindex its file database, making
it possible to the user to, for example, listen to new music without
having to rebooting the phone.

We believe this is what we all want for the platform.

Change-Id: I481d45320ac4a258eed13b91c487e536584b6579
Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>